MAJOR refactoring, namespacing, changing to the new notifications system
