+++ lib/active_support/cache/dalli_store_additions.rb
@@ -1,7 +1,6 @@
 require 'dalli'
 require 'digest/md5'
 require 'active_support/cache'
-require 'cashier/addons/plugins'
 
 module ActiveSupport
   module Cache
@@ -12,19 +11,19 @@ module ActiveSupport
     # an in memory cache inside of a block.
     class DalliStore < Store
       def write_with_tags(key, value, options = {})
-        Cashier::Addons::Plugins.call_plugin_method(:on_cache_write, key)
+        ActiveSupport::Notifications.instrument("cashier.write_cache_key", :data => key) do
+          tags = options.delete(:tag)
+          Cashier.store_fragment(key, tags) if tags
 
-        tags = options.delete(:tag)
-        Cashier.store_fragment(key, tags) if tags
-
-        write_without_tags(key, value, options)
+          write_without_tags(key, value, options)
+        end
       end
       alias_method_chain :write, :tags
 
-
       def delete_with_tags(key, options = nil)
-        Cashier::Addons::Plugins.call_plugin_method(:on_cache_delete, key)
-        delete_without_tags(key, options)
+        ActiveSupport::Notifications.instrument("cashier.delete_cache_key", :data => key) do
+          delete_without_tags(key, options)  
+        end
       end
 
       alias_method_chain :delete, :tags
