+++ lib/cashier.rb
@@ -1,10 +1,11 @@
 module Cashier
+
   extend self
 
   CACHE_KEY = 'cashier-tags'
 
   def adapter
-    Cashier::Addons::Adapters.current_adapter
+    Cashier::StoreAdapters.current_adapter
   end
 
   # Public: whether the module will perform caching or not. this is being set in the application layer .perform_caching configuration
@@ -29,15 +30,16 @@ module Cashier
   #
   def store_fragment(fragment, *tags)
     return unless perform_caching?
-    Cashier::Addons::Plugins.call_plugin_method(:on_store_fragment, [fragment, tags])
 
-    tags.each do |tag|
-      # store the fragment
-      adapter.store_fragment_in_tag(fragment, tag)
+    ActiveSupport::Notifications.instrument("cashier.store_fragment", :data => [fragment, tags]) do
+      tags.each do |tag|
+        # store the fragment
+        adapter.store_fragment_in_tag(fragment, tag)
+      end
+
+       # now store the tag for book keeping
+      adapter.store_tags(tags)
     end
-
-     # now store the tag for book keeping
-    adapter.store_tags(tags)
   end
 
   # Public: expire tags. expiring the keys 'assigned' to the tags you expire and removes the tags from the tags list
@@ -50,22 +52,23 @@ module Cashier
   # 
   def expire(*tags)
     return unless perform_caching?
-    Cashier::Addons::Plugins.call_plugin_method(:on_expire, tags)
 
-    # delete them from the cache
-    tags.each do |tag|
-      fragment_keys = adapter.get_fragments_for_tag(tag)
-      
-      fragment_keys.each do |fragment_key|
-        Rails.cache.delete(fragment_key)
+    ActiveSupport::Notifications.instrument("cashier.expire", :data => tags) do
+      # delete them from the cache
+      tags.each do |tag|
+        fragment_keys = adapter.get_fragments_for_tag(tag)
+        
+        fragment_keys.each do |fragment_key|
+          Rails.cache.delete(fragment_key)
+        end
+
+        adapter.delete_tag(tag)
       end
 
-      adapter.delete_tag(tag)
+      # now remove them from the list
+      # of stored tags
+      adapter.remove_tags(tags)
     end
-
-    # now remove them from the list
-    # of stored tags
-    adapter.remove_tags(tags)
   end
 
   # Public: returns the array of tags stored in the tags store.
@@ -88,8 +91,9 @@ module Cashier
   #   Cashier.clear
   #
   def clear
-    Cashier::Addons::Plugins.call_plugin_method(:on_clear)
-    adapter.clear
+    ActiveSupport::Notifications.instrument("cashier.clear") do
+      adapter.clear
+    end
   end
 
   # Public: get all the keys names as an array.
@@ -119,8 +123,7 @@ end
 
 
 require 'rails'
-require 'cashier/addons/plugins'
-require 'cashier/addons/adapters'
+require 'cashier/store_adapters'
 require 'active_support/cache/dalli_store_additions'
 require 'cashier/adapters/cache_store'
 require 'cashier/adapters/redis_store'
